order-api/requirements.txt
aniso8601==3.0.0
astroid==1.6.3
autopep8==1.3.5
click==6.7
docutils==0.14
Flask==0.12.2
flask-restplus==0.10.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kafka-python==1.4.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.4.0
pylint==1.8.4
python-dateutil==2.6.1
pytz==2018.4
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
